feat: Correciones V2
